// src/app/api/generated/model/authorized-context-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Contexto de trabajo habilitado para la cuenta autenticada
 */
export interface AuthorizedContextResponse { 
    /**
     * Identificador del consultorio
     */
    consultorioId?: number;
    /**
     * Nombre del consultorio
     */
    consultorioName?: string;
    /**
     * Identificador de la organizacion
     */
    organizationId?: number;
    /**
     * Nombre de la organizacion
     */
    organizationName?: string;
}